+++ src/letter/letter.controller.ts
@@ -13,7 +13,7 @@ import { SendLetterDto } from './dtos/send-letter.dto';
 import { SaveWritingDto } from './dtos/save-writing.dto';
 import { ApiTags } from '@nestjs/swagger';
 import { ApiLetters } from './letter.swagger';
-import { DeleteLettersDto } from './dtos/delete.letter.dto';
+import { DeleteLettersDto } from './dtos/delete-letter.dto';
 import { FirebaseAuthGuard } from 'src/auth/firebase-auth.guard';
 
 @Controller('letters')
